requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
pandas-ta>=0.3.14b
mplfinance>=0.12.10b0
numpy<2.0
matplotlib>=3.7.0
python-docx>=1.0.0
python-pptx>=0.6.21
